Type Alias pallet_pool_system::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn set_max_reserve( origin: OriginFor<T>, pool_id: T::PoolId, max_reserve: T::Balance ) -> DispatchResult

Sets the maximum reserve for a pool

The caller must have the LiquidityAdmin role in order to invoke this extrinsic. This role is not given to the pool creator by default, and must be added with the Permissions pallet before this extrinsic can be called.

pub fn close_epoch( origin: OriginFor<T>, pool_id: T::PoolId ) -> DispatchResultWithPostInfo

Close the current epoch

Closing an epoch locks in all invest and redeem orders placed during the epoch, and causes all further invest and redeem orders to be set for the next epoch.

If all orders can be executed without violating any pool constraints - which include maximum reserve and the tranche risk buffers - the execution will also be done. See execute_epoch for details on epoch execution.

If pool constraints would be violated by executing all orders, the pool enters a submission period. During a submission period, partial executions can be submitted to be scored, and the best-scoring solution will eventually be executed. See submit_solution.

pub fn submit_solution( origin: OriginFor<T>, pool_id: T::PoolId, solution: Vec<TrancheSolution> ) -> DispatchResultWithPostInfo

Submit a partial execution solution for a closed epoch

If the submitted solution is โbetterโ than the previous best solution, it will replace it. Solutions are ordered such that solutions which do not violate constraints are better than those that do.

Once a valid solution has been submitted, the challenge time begins. The pool can be executed once the challenge time has expired.

pub fn execute_epoch( origin: OriginFor<T>, pool_id: T::PoolId ) -> DispatchResultWithPostInfo

Execute an epoch for which a valid solution has been submitted.

Mints or burns tranche tokens based on investments and redemptions

Updates the portion of the reserve and loan balance assigned to each tranche, based on the investments and redemptions to those tranches.

impl<T: Config> Pallet<T>

pub fn score_solution( pool_id: &PoolDetailsOf<T>, epoch: &EpochExecutionInfo<<T as Config>::Balance, <T as Config>::BalanceRatio, <T as Config>::EpochId, <T as Config>::TrancheWeight, <T as Config>::BlockNumber, <T as Config>::TrancheCurrency, <T as Config>::MaxTranches>, solution: &[TrancheSolution] ) -> Result<EpochSolution<T::Balance, T::MaxTranches>, DispatchError>

Scores a solution.

This function checks the state a pool would be in when applying a solution to an epoch. Depending on the state, the correct scoring function is chosen.

pub fn is_valid_tranche_change( old_tranches: Option<&TranchesOf<T>>, new_tranches: &Vec<TrancheUpdate<T::Rate>> ) -> DispatchResult

impl<T: Config> Pallet<T>

pub fn pool<KArg>(k: KArg) -> Option<PoolDetailsOf<T>>where KArg: EncodeLike<T::PoolId>,

impl<T: Config> Pallet<T>

pub fn scheduled_update<KArg>( k: KArg ) -> Option<ScheduledUpdateDetails<<T as Config>::Rate, <T as Config>::MaxTokenNameLength, <T as Config>::MaxTokenSymbolLength, <T as Config>::MaxTranches>>where KArg: EncodeLike<T::PoolId>,

impl<T: Config> Pallet<T>

pub fn epoch_targets<KArg>( k: KArg ) -> Option<EpochExecutionInfo<<T as Config>::Balance, <T as Config>::BalanceRatio, <T as Config>::EpochId, <T as Config>::TrancheWeight, <T as Config>::BlockNumber, <T as Config>::TrancheCurrency, <T as Config>::MaxTranches>>where KArg: EncodeLike<T::PoolId>,

impl<T: Config> Pallet<T>

pub fn account_deposits<KArg>(k: KArg) -> T::Balancewhere KArg: EncodeLike<T::AccountId>,

impl<T: Config> Pallet<T>

pub fn pool_deposits<KArg>( k: KArg ) -> Option<PoolDepositInfo<<T as Config>::AccountId, <T as Config>::Balance>>where KArg: EncodeLike<T::PoolId>,