Type Alias pallet_pool_system::TrancheOf

pub type TrancheOf<T> = Tranche<<T as Config>::Balance, <T as Config>::Rate, <T as Config>::TrancheWeight, <T as Config>::TrancheCurrency>;

Types alias for Tranche

Aliased Type

struct TrancheOf<T> {
    pub tranche_type: TrancheType<<T as Config>::Rate>,
    pub seniority: u32,
    pub currency: <T as Config>::TrancheCurrency,
    pub debt: <T as Config>::Balance,
    pub reserve: <T as Config>::Balance,
    pub loss: <T as Config>::Balance,
    pub ratio: Perquintill,
    pub last_updated_interest: u64,
    pub _phantom: PhantomData<<T as Config>::TrancheWeight>,
}

Fields
tranche_type: TrancheType<<T as Config>::Rate>seniority: u32currency: <T as Config>::TrancheCurrencydebt: <T as Config>::Balancereserve: <T as Config>::Balanceloss: <T as Config>::Balanceratio: Perquintilllast_updated_interest: u64_phantom: PhantomData<<T as Config>::TrancheWeight>
Implementations

impl<Balance, Rate, Weight, Currency> Tranche<Balance, Rate, Weight, Currency>where Balance: Copy + BaseArithmetic + FixedPointOperand + Unsigned + From<u64> + MultiplyRational, Rate: FixedPointNumber<Inner = Balance> + One + Copy, Weight: Copy + From<u128>,

pub fn balance(&self) -> Result<Balance, ArithmeticError>

Returns the sum of the debt and reserve amounts.

pub fn free_balance(&self) -> Result<Balance, ArithmeticError>

Returns the reserve amount.

pub fn accrue(&mut self, now: Seconds) -> Result<(), ArithmeticError>

Update the debt of a Tranche by multiplying with the accrued interest since the last update: debt = debt * interest_rate_per_second ^ (now - last_update)

pub fn min_risk_buffer(&self) -> Perquintill

Returns the min risk buffer of a non-residual Tranche or zero.

pub fn interest_rate_per_sec(&self) -> Rate

Returns the interest rate per second for a non-residual Tranche or one.

pub fn debt(&mut self, now: Seconds) -> Result<Balance, DispatchError>

Updates the debt by applying the accrued interest rate since the last update moment and returns it.

pub fn create_asset_metadata( &self, decimals: u32, token_name: Vec<u8>, token_symbol: Vec<u8> ) -> AssetMetadata<Balance, CustomMetadata>where Balance: Zero, Currency: Encode, CustomMetadata: Parameter + Member + TypeInfo,