Type Alias pallet_pool_system::TranchesOf

pub type TranchesOf<T> = Tranches<<T as Config>::Balance, <T as Config>::Rate, <T as Config>::TrancheWeight, <T as Config>::TrancheCurrency, <T as Config>::TrancheId, <T as Config>::PoolId, <T as Config>::MaxTranches>;

Types alias for Tranches

Aliased Type

struct TranchesOf<T> {
    pub tranches: BoundedVec<Tranche<<T as Config>::Balance, <T as Config>::Rate, <T as Config>::TrancheWeight, <T as Config>::TrancheCurrency>, <T as Config>::MaxTranches>,
    pub ids: BoundedVec<<T as Config>::TrancheId, <T as Config>::MaxTranches>,
    pub salt: (u64, <T as Config>::PoolId),
}

Fields
tranches: BoundedVec<Tranche<<T as Config>::Balance, <T as Config>::Rate, <T as Config>::TrancheWeight, <T as Config>::TrancheCurrency>, <T as Config>::MaxTranches>ids: BoundedVec<<T as Config>::TrancheId, <T as Config>::MaxTranches>salt: (u64, <T as Config>::PoolId)
Implementations

impl<Balance, Rate, Weight, TrancheCurrency, TrancheId, PoolId, MaxTranches> Tranches<Balance, Rate, Weight, TrancheCurrency, TrancheId, PoolId, MaxTranches>where TrancheCurrency: Copy + TrancheCurrencyT<PoolId, TrancheId>, Balance: Zero + Copy + BaseArithmetic + FixedPointOperand + Unsigned + From<u64> + MultiplyRational, Weight: Copy + From<u128>, Rate: One + Copy + FixedPointNumber<Inner = Balance>, TrancheId: Clone + From<[u8; 16]> + PartialEq, PoolId: Copy + Encode, MaxTranches: Get<u32>,

pub fn from_input<MaxTokenNameLength, MaxTokenSymbolLength>( pool: PoolId, tranche_inputs: Vec<TrancheInput<Rate, MaxTokenNameLength, MaxTokenSymbolLength>>, now: Seconds ) -> Result<Self, DispatchError>where MaxTokenNameLength: Get<u32>, MaxTokenSymbolLength: Get<u32>,

pub fn tranche_currency( &self, id: TrancheLoc<TrancheId> ) -> Option<TrancheCurrency>

pub fn tranche_id(&self, id: TrancheLoc<TrancheId>) -> Option<TrancheId>

pub fn tranche_index(&self, id: &TrancheLoc<TrancheId>) -> Option<TrancheIndex>

pub fn get_mut_tranche( &mut self, id: TrancheLoc<TrancheId> ) -> Option<&mut Tranche<Balance, Rate, Weight, TrancheCurrency>>

pub fn get_tranche( &self, id: TrancheLoc<TrancheId> ) -> Option<&Tranche<Balance, Rate, Weight, TrancheCurrency>>

pub fn replace<MaxTokenNameLength, MaxTokenSymbolLength>( &mut self, at: TrancheIndex, tranche: TrancheInput<Rate, MaxTokenNameLength, MaxTokenSymbolLength>, now: Seconds ) -> DispatchResultwhere MaxTokenNameLength: Get<u32>, MaxTokenSymbolLength: Get<u32>,

pub fn validate_insert( &self, at: TrancheIndex, tranche: &Tranche<Balance, Rate, Weight, TrancheCurrency> ) -> DispatchResult

pub fn add<MaxTokenNameLength, MaxTokenSymbolLength>( &mut self, at: TrancheIndex, tranche: TrancheInput<Rate, MaxTokenNameLength, MaxTokenSymbolLength>, now: Seconds ) -> DispatchResultwhere MaxTokenNameLength: Get<u32>, MaxTokenSymbolLength: Get<u32>,

pub fn remove(&mut self, at: TrancheIndex) -> DispatchResult

Removing should only be possible if the Tranche at the given index has zero balance and is not the residual one.

pub fn ids_non_residual_top(&self) -> Vec<TrancheId>

pub fn of_pool(&self) -> PoolId

pub fn combine_non_residual_top<R, F>( &self, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&Tranche<Balance, Rate, Weight, TrancheCurrency>) -> Result<R, DispatchError>,

pub fn combine_mut_non_residual_top<R, F>( &mut self, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&mut Tranche<Balance, Rate, Weight, TrancheCurrency>) -> Result<R, DispatchError>,

pub fn combine_with_non_residual_top<R, I, W, F>( &self, with: I, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&Tranche<Balance, Rate, Weight, TrancheCurrency>, W) -> Result<R, DispatchError>, I: IntoIterator<Item = W>,

pub fn combine_with_mut_non_residual_top<R, W, I, F>( &mut self, with: I, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&mut Tranche<Balance, Rate, Weight, TrancheCurrency>, W) -> Result<R, DispatchError>, I: IntoIterator<Item = W>,

pub fn ids_residual_top(&self) -> Vec<TrancheId>

pub fn combine_residual_top<R, F>(&self, f: F) -> Result<Vec<R>, DispatchError>where F: FnMut(&Tranche<Balance, Rate, Weight, TrancheCurrency>) -> Result<R, DispatchError>,

pub fn combine_mut_residual_top<R, F>( &mut self, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&mut Tranche<Balance, Rate, Weight, TrancheCurrency>) -> Result<R, DispatchError>,

pub fn combine_with_residual_top<R, I, W, F>( &self, with: I, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&Tranche<Balance, Rate, Weight, TrancheCurrency>, W) -> Result<R, DispatchError>, I: IntoIterator<Item = W>,

pub fn combine_with_mut_residual_top<R, W, I, F>( &mut self, with: I, f: F ) -> Result<Vec<R>, DispatchError>where F: FnMut(&mut Tranche<Balance, Rate, Weight, TrancheCurrency>, W) -> Result<R, DispatchError>, I: IntoIterator<Item = W>,

pub fn calculate_prices<BalanceRatio, Tokens, AccountId>( &mut self, total_assets: Balance, now: Seconds ) -> Result<Vec<BalanceRatio>, DispatchError>where BalanceRatio: FixedPointNumber<Inner = Balance>, Tokens: Inspect<AccountId, Balance = Balance>, TrancheCurrency: Into<<Tokens as Inspect<AccountId>>::AssetId>,

Returns the current prices of the tranches based on the current NAV and each tranche’s balance and total issuance at this exact moment. The correctness of the waterfall is ensured by starting at the top non-residual tranch.

pub fn num_tranches(&self) -> usize

pub fn into_tranches( self ) -> BoundedVec<Tranche<Balance, Rate, Weight, TrancheCurrency>, MaxTranches>

pub fn non_residual_tranches( &self ) -> Option<&[Tranche<Balance, Rate, Weight, TrancheCurrency>]>

pub fn non_residual_tranches_mut( &mut self ) -> Option<&mut [Tranche<Balance, Rate, Weight, TrancheCurrency>]>

pub fn residual_tranche( &self ) -> Option<&Tranche<Balance, Rate, Weight, TrancheCurrency>>

pub fn residual_tranche_mut( &mut self ) -> Option<&mut Tranche<Balance, Rate, Weight, TrancheCurrency>>

pub fn non_residual_top_slice( &self ) -> &RevSlice<Tranche<Balance, Rate, Weight, TrancheCurrency>>

pub fn non_residual_top_slice_mut( &mut self ) -> &mut RevSlice<Tranche<Balance, Rate, Weight, TrancheCurrency>>

pub fn residual_top_slice( &self ) -> &[Tranche<Balance, Rate, Weight, TrancheCurrency>]

pub fn residual_top_slice_mut( &mut self ) -> &mut [Tranche<Balance, Rate, Weight, TrancheCurrency>]

pub fn supplies(&self) -> Result<Vec<Balance>, DispatchError>

Returns each tranche’s total supply starting at the residual top.

pub fn acc_supply(&self) -> Result<Balance, DispatchError>

Returns the total supply over all tranches starting at the residual top.

pub fn min_risk_buffers(&self) -> Vec<Perquintill>

Returns each tranche’s min risk buffer starting at the residual top.

pub fn seniorities(&self) -> Vec<Seniority>

Returns each tranche’s seniority starting at the residual top.

pub fn rebalance_tranches( &mut self, now: Seconds, pool_total_reserve: Balance, pool_nav: Balance, tranche_ratios: &[Perquintill], executed_amounts: &[(Balance, Balance)] ) -> DispatchResult